updated to last of 3.5

+++ examples/Integer/probable_primroot.C
@@ -29,7 +29,7 @@ int main(int argc, char** argv)
 {
     IntNumTheoDom<> IP;
 #ifdef __GMP_PLUSPLUS__
-    IP.seeding( (unsigned long)BaseTimer::seed() );
+    IP.seeding( BaseTimer::seed() );
 #endif
 
     double error;
